Canvas.py: Fix the bug that caused by float font size
git-svn-id: https://ubuntu-tweak.googlecode.com/svn/trunk@325 89e9d5ea-4437-0410-aedb-dd2c2f29ea1f

+++ src/common/Canvas.py
@@ -12,7 +12,7 @@ from Settings import StringSetting
 
 class RenderCell:
     font_name = StringSetting('/desktop/gnome/interface/font_name').get_string()
-    font_size = int(font_name.split()[-1])
+    font_size = float(font_name.split()[-1])
 
     def __init__(self, 
                 cell = None,
